requirements.txt
# --- Core FastAPI stack ---
fastapi>=0.109.0
uvicorn[standard]>=0.22.0
gunicorn

# --- Database + ORM ---
SQLAlchemy>=2.0.30
alembic>=1.13.2
databases>=0.8.0          # optional async DB toolkit

# --- Data validation ---
pydantic>=2.7.0

# --- Environment management ---
python-dotenv>=1.0.1

# --- Utilities ---
requests>=2.31.0
python-dateutil
# --- App ---
app
